pyproject.toml
[project]
name = "openadr3-client-gac-compliance"
version = "0.1.0"
description = ""
authors = [
    {name = "Nick van der Burgt", email = "nick.van.der.burgt@elaad.nl"}
]
readme = "README.md"
requires-python = ">=3.12, <4"
dependencies = [
    "pydantic (>=2.11.2,<3.0.0)",
    "openadr3-client (>=0.0.11,<1.0.0)",
    "pycountry (>=24.6.1,<25.0.0)",
    "python-decouple (>=3.8,<4.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.11.4"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"

[[tool.mypy.overrides]]
module = ["openadr3_client", "decouple"]
ignore_missing_imports = true

[tool.taskipy.tasks]
local-ci = "task check && task format && task mypy && task test"
fix = "task check-fix && task format-fix && task mypy"
check = "ruff check ."
check-fix = "ruff check --fix ."
check-ci = "ruff check --output-format=github --no-cache ."
mypy = "mypy . --check-untyped-defs"
format = "ruff format . --diff"
format-fix = "ruff format ."
format-ci = "ruff format --diff --no-cache ."
test = "pytest --cov"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]

# Do not require docstrings for functions, methods and classes
ignore = [
    # From:https://github.com/wemake-services/wemake-django-template/blob/master/pyproject.toml
    "A005",   # allow to shadow stdlib and builtin module names
    "COM812", # trailing comma, conflicts with `ruff format`
    # Different doc rules that we don't really care about:
    "D100",
    "D104",
    "D106",
    "D203",
    "D212",
    "D401",
    "D404",
    "D405",
    "ISC001",  # implicit string concat conflicts with `ruff format`
    "ISC003",  # prefer explicit string concat over implicit concat
    "PLR09",   # we have our own complexity rules
    # "PLR2004", # do not report magic numbers -> I do want to report magic numbers
    "PLR6301", # do not require classmethod / staticmethod when self not used
    "TRY003",  # long exception messages from `tryceratops`
    # Start of own error codes
    "ANN002", # Ignore unused *args
    "ANN003", # Ignore unused **kwargs
    "D107", # Allow undocumented __init__
    "TC001", # Move to type checking block (does not play nice with pydantic models)
    "TC002", # Move to type checking block (does not play nice with pydantic models)
]

# Ignores S101 ("Use of 'assert' detected") for the tests.
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN201", "SLF001", "S106", "S105", "S311", "PLR2004", "PT018"]